main/backend/app/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
pymysql==1.1.0
Flask-Migrate==4.0.7
Flask-Session==0.7.0

cryptography
pydantic
boto3
python-jose
requests
